Refactor compatibility generation
