--- pkg/config/cuda_base_image_tags.json
@@ -1,36 +1,0 @@
-[
-  "12.1.1-cudnn8-devel-ubuntu22.04",
-  "12.1.1-cudnn8-devel-ubuntu20.04",
-  "12.1.0-cudnn8-devel-ubuntu22.04",
-  "12.1.0-cudnn8-devel-ubuntu20.04",
-  "12.0.1-cudnn8-devel-ubuntu22.04",
-  "12.0.1-cudnn8-devel-ubuntu20.04",
-  "12.0.1-cudnn8-devel-ubuntu18.04",
-  "12.0.0-cudnn8-devel-ubuntu22.04",
-  "12.0.0-cudnn8-devel-ubuntu20.04",
-  "12.0.0-cudnn8-devel-ubuntu18.04",
-  "11.8.0-cudnn8-devel-ubuntu22.04",
-  "11.8.0-cudnn8-devel-ubuntu20.04",
-  "11.8.0-cudnn8-devel-ubuntu18.04",
-  "11.7.1-cudnn8-devel-ubuntu22.04",
-  "11.7.1-cudnn8-devel-ubuntu20.04",
-  "11.7.1-cudnn8-devel-ubuntu18.04",
-  "11.6.2-cudnn8-devel-ubuntu20.04",
-  "11.6.2-cudnn8-devel-ubuntu18.04",
-  "11.5.2-cudnn8-devel-ubuntu20.04",
-  "11.5.2-cudnn8-devel-ubuntu18.04",
-  "11.4.3-cudnn8-devel-ubuntu20.04",
-  "11.4.3-cudnn8-devel-ubuntu18.04",
-  "11.3.1-cudnn8-devel-ubuntu20.04",
-  "11.3.1-cudnn8-devel-ubuntu18.04",
-  "11.3.1-cudnn8-devel-ubuntu16.04",
-  "11.2.2-cudnn8-devel-ubuntu20.04",
-  "11.2.2-cudnn8-devel-ubuntu18.04",
-  "11.2.2-cudnn8-devel-ubuntu16.04",
-  "11.1.1-cudnn8-devel-ubuntu20.04",
-  "11.1.1-cudnn8-devel-ubuntu18.04",
-  "11.1.1-cudnn8-devel-ubuntu16.04",
-  "11.0.3-cudnn8-devel-ubuntu20.04",
-  "11.0.3-cudnn8-devel-ubuntu18.04",
-  "11.0.3-cudnn8-devel-ubuntu16.04"
-]
