+++ tools/compatgen/internal/torch.go
@@ -1,187 +1,15 @@
-package main
+package internal
 
 import (
-	"encoding/json"
-	"flag"
 	"fmt"
-	"os"
 	"regexp"
-	"sort"
-	"strconv"
 	"strings"
 
 	"github.com/anaskhan96/soup"
 
 	"github.com/hashicorp/go-version"
 	"github.com/replicate/cog/pkg/config"
-	"github.com/replicate/cog/pkg/util/console"
 )
-
-func main() {
-	tfOutputPath := flag.String("tf-output", "pkg/config/tf_compatability_matrix.json", "Tensorflow output path")
-	torchOutputPath := flag.String("torch-output", "pkg/config/torch_compatability_matrix.json", "PyTorch output path")
-	cudaImagesOutputPath := flag.String("cuda-images-output", "pkg/config/cuda_base_image_tags.json", "CUDA base images output path")
-	flag.Parse()
-
-	if *tfOutputPath == "" && *torchOutputPath == "" && *cudaImagesOutputPath == "" {
-		console.Fatal("at least one of -tf-output, -torch-output, -cuda-images-output must be provided")
-	}
-
-	if *tfOutputPath != "" {
-		if err := writeTFCompatibilityMatrix(*tfOutputPath); err != nil {
-			console.Fatalf("Failed to write Tensorflow compatibility matrix: %s", err)
-		}
-	}
-	if *torchOutputPath != "" {
-		if err := writeTorchCompatibilityMatrix(*torchOutputPath); err != nil {
-			console.Fatalf("Failed to write PyTorch compatibility matrix: %s", err)
-		}
-	}
-	if *cudaImagesOutputPath != "" {
-		if err := writeCUDABaseImageTags(*cudaImagesOutputPath); err != nil {
-			console.Fatalf("Failed to write CUDA base images: %s", err)
-		}
-	}
-}
-
-func writeTFCompatibilityMatrix(outputPath string) error {
-	console.Infof("Writing Tensorflow compatibility matrix to %s...", outputPath)
-
-	url := "https://www.tensorflow.org/install/source"
-	resp, err := soup.Get(url)
-	if err != nil {
-		return fmt.Errorf("Failed to download %s: %w", url, err)
-	}
-	doc := soup.HTMLParse(resp)
-	gpuHeading := doc.Find("h4", "id", "gpu")
-	table := gpuHeading.FindNextElementSibling()
-	rows := table.FindAll("tr")
-
-	compats := []config.TFCompatibility{}
-	for _, row := range rows[1:] {
-		cells := row.FindAll("td")
-		gpuPackage, packageVersion := split2(cells[0].Text(), "-")
-		pythonVersions, err := parsePythonVersionsCell(cells[1].Text())
-		if err != nil {
-			return err
-		}
-		cuDNN := cells[4].Text()
-		cuda := cells[5].Text()
-
-		compat := config.TFCompatibility{
-			TF:           packageVersion,
-			TFCPUPackage: "tensorflow==" + packageVersion,
-			TFGPUPackage: gpuPackage + "==" + packageVersion,
-			CUDA:         cuda,
-			CuDNN:        cuDNN,
-			Pythons:      pythonVersions,
-		}
-		compats = append(compats, compat)
-	}
-
-	// sanity check
-	if len(compats) < 21 {
-		return fmt.Errorf("Tensorflow compatibility matrix only had %d rows, has the html changed?", len(compats))
-	}
-
-	data, err := json.MarshalIndent(compats, "", "  ")
-	if err != nil {
-		return err
-	}
-	if err := os.WriteFile(outputPath, data, 0o644); err != nil {
-		return err
-	}
-	return nil
-}
-
-func writeTorchCompatibilityMatrix(outputPath string) error {
-	console.Infof("Writing PyTorch compatibility matrix to %s...", outputPath)
-
-	compats := []config.TorchCompatibility{}
-	var err error
-	compats, err = fetchCurrentTorchVersions(compats)
-	if err != nil {
-		return err
-	}
-	compats, err = fetchPreviousTorchVersions(compats)
-	if err != nil {
-		return err
-	}
-
-	// sanity check
-	if len(compats) < 21 {
-		return fmt.Errorf("PyTorch compatibility matrix only had %d rows, has the html changed?", len(compats))
-	}
-
-	data, err := json.MarshalIndent(compats, "", "  ")
-	if err != nil {
-		return err
-	}
-	if err := os.WriteFile(outputPath, data, 0o644); err != nil {
-		return err
-	}
-	return nil
-}
-
-func getCUDABaseImageTags(url string) ([]string, error) {
-	tags := []string{}
-
-	resp, err := soup.Get(url)
-	if err != nil {
-		return tags, fmt.Errorf("Failed to download %s: %w", url, err)
-	}
-
-	var results struct {
-		Next    *string
-		Results []struct {
-			Name string `json:"name"`
-		} `json:"results"`
-	}
-	if err := json.Unmarshal([]byte(resp), &results); err != nil {
-		return tags, fmt.Errorf("Failed parse CUDA images json: %w", err)
-	}
-
-	for _, result := range results.Results {
-		tag := result.Name
-		if strings.Contains(tag, "-cudnn") && !strings.HasSuffix(tag, "-rc") {
-			tags = append(tags, tag)
-		}
-	}
-
-	// recursive case for pagination
-	if results.Next != nil {
-		nextURL := *results.Next
-		nextTags, err := getCUDABaseImageTags(nextURL)
-		if err != nil {
-			return tags, err
-		}
-		tags = append(tags, nextTags...)
-	}
-
-	return tags, nil
-}
-
-func writeCUDABaseImageTags(outputPath string) error {
-	console.Infof("Writing CUDA base images to %s...", outputPath)
-	url := "https://hub.docker.com/v2/repositories/nvidia/cuda/tags/?page_size=1000&name=devel-ubuntu&ordering=last_updated"
-
-	tags, err := getCUDABaseImageTags(url)
-	if err != nil {
-		return err
-	}
-
-	sort.Sort(sort.Reverse(sort.StringSlice(tags)))
-
-	data, err := json.MarshalIndent(tags, "", "  ")
-	if err != nil {
-		return err
-	}
-	if err := os.WriteFile(outputPath, data, 0o644); err != nil {
-		return err
-	}
-
-	return nil
-}
 
 type torchPackage struct {
 	Name          string
@@ -189,6 +17,26 @@ type torchPackage struct {
 	Variant       string
 	CUDA          *string
 	PythonVersion string
+}
+
+func FetchTorchCompatibilityMatrix() ([]config.TorchCompatibility, error) {
+	compats := []config.TorchCompatibility{}
+	var err error
+	compats, err = fetchCurrentTorchVersions(compats)
+	if err != nil {
+		return nil, err
+	}
+	compats, err = fetchPreviousTorchVersions(compats)
+	if err != nil {
+		return nil, err
+	}
+
+	// sanity check
+	if len(compats) < 21 {
+		return nil, fmt.Errorf("PyTorch compatibility matrix only had %d rows, has the html changed?", len(compats))
+	}
+
+	return compats, nil
 }
 
 func fetchTorchPackages(name string) ([]torchPackage, error) {
@@ -453,56 +301,4 @@ func fixTorchCompatibility(compat *config.TorchCompatibility) {
 	if strings.HasPrefix(compat.Torchvision, "0.8.0") {
 		compat.Torchvision = strings.Replace(compat.Torchvision, "0.8.0", "0.8.1", -1)
 	}
-}
-
-func parsePythonVersionsCell(val string) ([]string, error) {
-	versions := []string{}
-	parts := strings.Split(val, ",")
-	for _, part := range parts {
-		part = strings.TrimSpace(part)
-		if strings.Contains(part, "-") {
-			start, end := split2(part, "-")
-			startMajor, startMinor, err := splitPythonVersion(start)
-			if err != nil {
-				return nil, err
-			}
-			endMajor, endMinor, err := splitPythonVersion(end)
-			if err != nil {
-				return nil, err
-			}
-
-			if startMajor != endMajor {
-				return nil, fmt.Errorf("Invalid start and end minor versions: %d, %d", startMajor, endMajor)
-			}
-			for minor := startMinor; minor <= endMinor; minor++ {
-				versions = append(versions, newVersion(startMajor, minor))
-			}
-		} else {
-			versions = append(versions, part)
-		}
-	}
-	return versions, nil
-}
-
-func newVersion(major int, minor int) string {
-	return fmt.Sprintf("%d.%d", major, minor)
-}
-
-func splitPythonVersion(version string) (major int, minor int, err error) {
-	version = strings.TrimSpace(version)
-	majorStr, minorStr := split2(version, ".")
-	major, err = strconv.Atoi(majorStr)
-	if err != nil {
-		return 0, 0, err
-	}
-	minor, err = strconv.Atoi(minorStr)
-	if err != nil {
-		return 0, 0, err
-	}
-	return major, minor, nil
-}
-
-func split2(s string, sep string) (string, string) {
-	parts := strings.SplitN(s, sep, 2)
-	return parts[0], parts[1]
 }
